api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
python-pptx==0.6.23
kokoro-onnx==0.3.9
soundfile==0.13.1
jinja2==3.1.3
tqdm==4.66.5
sqlalchemy
psycopg2-binary
sqladmin
asyncpg
greenlet
obsws-python==1.2.0
mutagen==1.46.0
moviepy==2.2.1
ruff>=0.0.337
black>=24.12.0
isort>=6.5.5